Fix «symbol definition is void» error on first start.
When disassembly buffer is first invalidated in `gdb-update`,
`gdb-main-file` or `gdb-selected-file` may be not set yet. By using
`gdb-get-buffer` as a demand predicate in `gdb-invalidate-disassembly`
trigger, we make sure that disassembly trigger is activated only when
the user explicitly starts the disassembly buffer. Additionally we
make sure that there's a file to disassemble by raising error in other
case.