Publisher/subscriber model for GDB buffers.
Triggers which update buffers are now stored in `gdb-buffer-rules`
(formerly known as `gdb-buffer-rules-assoc`). When a buffer is created
via `gdb-get-buffer-create`, if it has an associated trigger (checked
by `gdb-rules-update-trigger`), the buffer and its trigger are added
to `gdb-buf-publisher` list, which is used to notify its members that
updating is needed. As the result, we don't need to explicitly call
all the invalidation triggers in `gdb-update`.

Notifications are implemented using new `gdb-add-subscriber` and
`gdb-emit-signal` functions.

`gdb-get-buffer-create` now accepts an optional thread argument which
sets a buffer-local `gdb-thread-number` variable to that value.

Triggers are now called from the respective buffers (which allows
utilizing of buffer-local variables), as well as handlers and
name-making functions. Combined with new `gdb-thread-number` semantics
it allows to keep several buffers of the same type for different
threads. Triggers are «bound» (with `gdb-bind-function-to-buffer`) to
buffers in `gdb-get-buffer-create`. We need this to avoid having to
provide a buffer for which the trigger call is needed to the trigger
itself. Triggers must accept one argument which is the name of
notification signal (it's currently ignored anyways).

`def-gdb-auto-updated-buffer` is renamed to
`def-gdb-trigger-and-handler` to better reflect its semantics.

Handler defined by `def-gdb-auto-update-handler` now calls
`gdb-update-buffer-name` after custom-defun.

`gdb-set-buffer-rules` now accepts fourth argument to specify
buffer-updating trigger.

`def-gdb-display-buffer` and `def-gdb-frame-for-buffer` now define
functions with one optinal thread argument which is fed to
`gdb-get-buffer-create` under the hood.

`gdb-update-buffer-name` calls a name-maker function associated with
current buffer type in `gdb-buffer-rules`.

`gdb-pending-triggers` now contain not just trigger names, but pairs
of `(buffer . trigger-name)`.

`gdb-stack-buffer-name` now sets buffer name to contain thread number.